+++ NiLangV2.1/NiLangV2.1.vcxproj.filters
@@ -31,11 +31,6 @@
     <None Include="packages.config" />
   </ItemGroup>
   <ItemGroup>
-    <ClCompile Include="lexer.cpp">
-      <Filter>Source Files</Filter>
-    </ClCompile>
-  </ItemGroup>
-  <ItemGroup>
     <Bison Include="grammar.y">
       <Filter>Source Files</Filter>
     </Bison>
